requirements.txt
playwright
playwright-stealth
gspread
google-auth
pandas
